2181. Merge Nodes in between Zeros
Q. You are given the head of a linked list, which contains a series of integers separated by 0's. The beginning and end of the linked list will have Node.val == 0. For every two consecutive 0's, merge all the nodes lying in between them into a single node whose value is the sum of all the merged nodes. The modified list should not contain any 0's.
Return the head of the modified linked list.

Approach: Here, I have used a list 'a' to traverse list 'head'.
List 'dummy' is the one returned with the new values and 'b' is used to traverse it.
count is used to add the values upto the next zero.
when value is zero, b->next is added with the value count.
and count is made zero. it is iterated till a reaches the end.

Time Complexity: O(n)

Space complexity: O(n), where n is the number of segments formed by zeroes.