31. Next Permutation
Q. The next permutation of an array of integers is the next lexicographically greater permutation of its integer. More formally, if all the permutations of the array are sorted in one container according to their lexicographical order, then the next permutation of that array is the permutation that follows it in the sorted container. If such arrangement is not possible, the array must be rearranged as the lowest possible order (i.e., sorted in ascending order).

Intuition:
Find the first decreasing element from the right
Find the smallest element greater than nums[i] in the suffix
Swap nums[i] with this smallest greater element
Reverse the suffix (right side of i)

Approach:
Find i, the rightmost index where nums[i] < nums[i+1].
If no such i exists, the list is in descending order, so simply reverse the list to get the smallest permutation.
Find j, the smallest element in nums[i+1:] that is greater than nums[i], and swap nums[i] and nums[j].
Reverse the elements after index i.

Time and Space Complexity:
O(n) & O(1)